Created the DBCreateSQLite3 directory to initialize an SQLite3 database and made changes to DBCreatePostgres to make it more compatible with a multidatabase environment:
- ISO639: message in the generated output stating to keep in sync with SQLite3
- ONIX: message in the generated output stating to keep in sync with SQLite3
- TMK_eCommerce_Create_6_InitialData.sql: changed the use of the POSITION function
  for LIKE because SQLite3 does not support POSITION.
- TMK_eCommerce_Create_9_Other_CodeTables.sql: changed a few ORDER BYes to force
  same ordering on both databases.

As for the steps executed, some steps are not needed in SQLite3 but the script
has the steps anyway (reported as "N/A") to make easier to keep up with both
scripts. The Postgres case also has one not needed step (step 10) that is
required for SQLite3.

Some of the sql scripts (TMK_eCommerce_Create_*.sql) must be manualy kept in
sync between both dirs. For simplicity, some basic automated conversion is
applied to the scripts for SQLite3 usage (for example: convert true/false to 1/0,
manualy assign values to expressions like "nextval('sequence')", etc).